pyproject.toml
[tool.poetry]
name = "bingo_memes"
version = "0.1.0"
description = "A Django/React webapp."
authors = ["long"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2"
djangorestframework = "^3.12.4"
ipython = "^7.22.0"
neovim = "^0.3.1"
flake8 = "^3.9.2"
autopep8 = "^1.5.7"
django-create-react-app = "^0.8.4"
psycopg2 = "^2.9.1"
python-decouple = "^3.4"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
